requirements.txt
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-leaflet==1.0.15
geopandas
requests
gunicorn
fiona
shapely
plotly==5.24.1
Flask==3.0.3
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0


